Simplify sync logic between WYSIWYG and HTML editors
Basically, changes made to the focused editor, whether it is a WYSIWYG
or HTML editor, are immediately synced to the fieldStore and
subsequently to the unfocused editor. Exceptionally, when a note is
loaded (or reloaded), the content of the fieldStore is pushed to both
editors, regardless of whether they are focused or not.

Also, <anki-mathjax> tags will not now show up in HTML editor unless the
user types them in manually.